+++ src/main/java/com/example/hackathon6backend/domain/auth/service/LoginService.java
@@ -6,12 +6,12 @@ import com.example.hackathon6backend.domain.room.entity.ClassRoom;
 import com.example.hackathon6backend.domain.room.exception.ClassRoomNotFoundException;
 import com.example.hackathon6backend.domain.room.repository.ClassRoomRepository;
 import com.example.hackathon6backend.domain.user.dto.response.UserResponse;
-import com.example.hackathon6backend.domain.user.entity.Student;
-import com.example.hackathon6backend.domain.user.entity.Teacher;
+import com.example.hackathon6backend.domain.student.entity.Student;
+import com.example.hackathon6backend.domain.teacher.entity.Teacher;
 import com.example.hackathon6backend.domain.user.entity.User;
-import com.example.hackathon6backend.domain.user.entity.repository.StudentRepository;
-import com.example.hackathon6backend.domain.user.entity.repository.TeacherRepository;
-import com.example.hackathon6backend.domain.user.entity.repository.UserRepository;
+import com.example.hackathon6backend.domain.student.repository.StudentRepository;
+import com.example.hackathon6backend.domain.teacher.repository.TeacherRepository;
+import com.example.hackathon6backend.domain.user.repository.UserRepository;
 import com.example.hackathon6backend.domain.user.exception.InvalidRoleException;
 import com.example.hackathon6backend.domain.user.exception.PasswordMisMatchException;
 import com.example.hackathon6backend.global.security.jwt.JwtProperties;
