+++ src/main/java/com/example/hackathon6backend/domain/range/service/StudentRangeService.java
@@ -6,8 +6,8 @@ import com.example.hackathon6backend.domain.range.repository.RangeRepository;
 import com.example.hackathon6backend.domain.room.entity.ClassRoom;
 import com.example.hackathon6backend.domain.room.exception.ClassRoomNotFoundException;
 import com.example.hackathon6backend.domain.room.repository.ClassRoomRepository;
-import com.example.hackathon6backend.domain.user.entity.Student;
-import com.example.hackathon6backend.domain.user.entity.repository.StudentRepository;
+import com.example.hackathon6backend.domain.student.entity.Student;
+import com.example.hackathon6backend.domain.student.repository.StudentRepository;
 import com.example.hackathon6backend.global.error.exception.ErrorCode;
 import com.example.hackathon6backend.global.error.exception.HackathonException;
 import lombok.RequiredArgsConstructor;
