+++ src/main/java/com/example/hackathon6backend/domain/student/repository/StudentRepository.java
@@ -1,6 +1,6 @@
-package com.example.hackathon6backend.domain.user.entity.repository;
+package com.example.hackathon6backend.domain.student.repository;
 
-import com.example.hackathon6backend.domain.user.entity.Student;
+import com.example.hackathon6backend.domain.student.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentRepository extends JpaRepository<Student, Long> {
