+++ src/main/java/com/example/hackathon6backend/domain/subject/service/SubjectService.java
@@ -12,8 +12,8 @@ import com.example.hackathon6backend.domain.subject.entity.Subject;
 import com.example.hackathon6backend.domain.room.repository.ClassRoomRepository;
 import com.example.hackathon6backend.domain.subject.exception.SubjectNotFoundException;
 import com.example.hackathon6backend.domain.subject.repository.SubjectRepository;
-import com.example.hackathon6backend.domain.user.entity.Teacher;
-import com.example.hackathon6backend.domain.user.entity.repository.TeacherRepository;
+import com.example.hackathon6backend.domain.teacher.entity.Teacher;
+import com.example.hackathon6backend.domain.teacher.repository.TeacherRepository;
 import com.example.hackathon6backend.global.error.exception.ErrorCode;
 import com.example.hackathon6backend.global.error.exception.HackathonException;
 import lombok.RequiredArgsConstructor;
