+++ src/main/java/com/example/hackathon6backend/domain/teacher/repository/TeacherRepository.java
@@ -1,6 +1,6 @@
-package com.example.hackathon6backend.domain.user.entity.repository;
+package com.example.hackathon6backend.domain.teacher.repository;
 
-import com.example.hackathon6backend.domain.user.entity.Teacher;
+import com.example.hackathon6backend.domain.teacher.entity.Teacher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TeacherRepository extends JpaRepository<Teacher, Long> {
