--- src/main/java/com/example/hackathon6backend/domain/user/entity/Role.java
@@ -1,6 +1,0 @@
-package com.example.hackathon6backend.domain.user.entity;
-
-public enum Role {
-    SCH, // 선생님
-    STU // 학생
-}
