+++ src/main/java/com/example/hackathon6backend/domain/user/entity/role/Role.java
@@ -1,0 +1,6 @@
+package com.example.hackathon6backend.domain.user.entity.role;
+
+public enum Role {
+    SCH, // 선생님
+    STU // 학생
+}
